+++ extension/js/buildShuffleButton.js
@@ -77,14 +77,14 @@ function buildShuffleButton(pageType) {
 		// Unhide the button if it was hidden
 		document.getElementById(buttonDivID).style.display = "flex";
 		if(mustUpdateVideoId) {
-			document.getElementById('youtube-random-video-shuffle-button-channel').children[0].children[0].children[0].children.namedItem('videoLink').innerHTML = videoId;
+			document.getElementById(buttonDivID).children[0].children[0].children[0].children.namedItem('videoLink').innerHTML = videoId;
 		}
 
 		return;
 	}
 
 	if (!currentChannel || !videoId) {
-		console.log("Cannot build button: Not on a channel page, or no video ID to identify channel available.");
+		console.log("Cannot build button: Not on a channel page, or no video ID to identify channel.");
 		return;
 	}