+++ extension/js/popup.js
@@ -2,47 +2,73 @@ const defaultApiKey = await chrome.runtime.sendMessage({ command: "getDefaultApi
 
 let configSync = await fetchConfigSync();
 
-// ---------- Get DOM elements ----------
+// ---------- Get relevant DOM elements ----------
 
 const domElements = {
+	// Custom API key: Option toggle
 	useCustomApiKeyOptionToggle: document.getElementById("useCustomApiKeyOptionToggle"),
-	dbSharingOptionToggle: document.getElementById("dbSharingOptionToggle"),
+	// Custom API key: Input
 	customApiKeyInputDiv: document.getElementById("customApiKeyInputDiv"),
 	customApiKeyInputField: customApiKeyInputDiv.children.namedItem("customApiKeyInputField"),
 	customApiKeySubmitButton: customApiKeyInputDiv.children.namedItem("customApiKeySubmitButton"),
 	customApiKeyInputErrorDiv: customApiKeyInputDiv.children.namedItem("customApiKeyInputErrorDiv"),
 	customApiKeyInputErrorText: customApiKeyInputErrorDiv.children.namedItem("customApiKeyInputErrorText"),
+	// Database sharing: Option toggle
+	dbSharingOptionToggle: document.getElementById("dbSharingOptionToggle"),
+	// Shuffling: Open in new tab option toggle
+	shuffleOpenInNewTabOptionToggle: document.getElementById("shuffleOpenInNewTabOptionToggle"),
+	// Shuffling: Open as playlist option toggle
+	shuffleOpenAsPlaylistOptionToggle: document.getElementById("shuffleOpenAsPlaylistOptionToggle"),
+	// Shuffling: Shuffle from last x% of videos input
+	shuffleLastXVideosInputField: document.getElementById("shuffleLastXVideosInputField"),
 }
 
 // ---------- Set default values from config ----------
 
 function setDomElementDefaultsFromConfig() {
+	// ----- Custom API key: Option toggle -----
 	// If this option is checked is only dependent on the value in sync storage
 	domElements.useCustomApiKeyOptionToggle.checked = configSync.useCustomApiKeyOption;
+
+	// ----- Database sharing: Option toggle -----
 	// Determine if the dbSharingOptionToggle should be checked and enabled
 	manageDbOptOutOption();
+
+	// ----- Custom API key: Input -----
 	// Show the customAPIKeyInputDiv if the user has enabled the option
 	if (configSync.useCustomApiKeyOption) {
 		domElements.customApiKeyInputDiv.classList.remove("hidden");
 	}
 	// Set the value of the custom API key input field to the value in sync storage
 	domElements.customApiKeyInputField.value = configSync.customYoutubeApiKey ? configSync.customYoutubeApiKey : "";
+
+	// ----- Shuffling: Open in new tab option toggle -----
+	domElements.shuffleOpenInNewTabOptionToggle.checked = configSync.shuffleOpenInNewTabOption;
+
+	// ----- Shuffling: Open as playlist option toggle -----
+	domElements.shuffleOpenAsPlaylistOptionToggle.checked = configSync.shuffleOpenAsPlaylistOption;
+
+	// ----- Shuffling: Shuffle from last x% of videos input -----
+	domElements.shuffleLastXVideosInputField.value = configSync.shuffleLastXVideosPercentage;
 }
 
 setDomElementDefaultsFromConfig();
 
 // ---------- Event listeners ----------
 
+// Custom API key: Option toggle
 domElements.useCustomApiKeyOptionToggle.addEventListener("change", function () {
 	setSyncStorageValue("useCustomApiKeyOption", this.checked);
-	manageDependents(useCustomApiKeyOptionToggle, this.checked);
+	manageDependents(domElements.useCustomApiKeyOptionToggle, this.checked);
 });
 
+// Database sharing: Option toggle
 domElements.dbSharingOptionToggle.addEventListener("change", function () {
 	setSyncStorageValue("databaseSharingEnabledOption", this.checked);
-	manageDependents(dbSharingOptionToggle, this.checked);
+	manageDependents(domElements.dbSharingOptionToggle, this.checked);
 });
 
+// Custom API key: Input
 domElements.customApiKeySubmitButton.addEventListener("click", async function () {
 	// Make sure the passed API key is valid
 	const newApiKey = domElements.customApiKeyInputField.value;
@@ -56,8 +82,36 @@ domElements.customApiKeySubmitButton.addEventListener("click", async function ()
 	manageDbOptOutOption();
 });
 
+// Shuffling: Open in new tab option toggle
+domElements.shuffleOpenInNewTabOptionToggle.addEventListener("change", function () {
+	setSyncStorageValue("shuffleOpenInNewTabOption", this.checked);
+	manageDependents(domElements.shuffleOpenInNewTabOptionToggle, this.checked);
+});
+
+// Shuffling: Open as playlist option toggle
+domElements.shuffleOpenAsPlaylistOptionToggle.addEventListener("change", function () {
+	setSyncStorageValue("shuffleOpenAsPlaylistOption", this.checked);
+	manageDependents(domElements.shuffleOpenAsPlaylistOptionToggle, this.checked);
+});
+
+// Shuffling: Shuffle from last x% of videos input
+domElements.shuffleLastXVideosInputField.addEventListener("focusout", function () {
+	// Clamp the value to the range [1, 100]
+	if (this.value === "") {
+		this.value = 100;
+	}
+	const value = Math.min(Math.max(this.value, 1), 100);
+	setSyncStorageValue("shuffleLastXVideosPercentage", value);
+	// Set the value of the input field to the clamped value
+	this.value = value;
+	manageDependents(domElements.shuffleLastXVideosInputField, value);
+});
+
+// ----- Dependency management -----
+
 function manageDependents(parent, checked) {
 	switch (parent) {
+		// Custom API key: Option toggle
 		case domElements.useCustomApiKeyOptionToggle:
 			if (checked) {
 				// Show input field for custom API key